requirements.txt
flask
gunicorn
google-cloud-speech
soundfile